LayoutTests/fast/box-shadow/box-shadow-parsing-invalid-expected.txt

Test the parsing of box-shadow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Negative blur radius value : 10px 10px -1px rgb(255, 0, 0)
PASS cssRule.type is 1
PASS declaration.length is 0
PASS declaration.getPropertyValue('box-shadow') is null

Negative blur radius value, with a spread defined : 10px 10px -1px 10px rgb(255, 0, 0)
PASS cssRule.type is 1
PASS declaration.length is 0
PASS declaration.getPropertyValue('box-shadow') is null

Negative blur radius value, with a negative spread defined : 10px 10px -1px -1px rgb(255, 0, 0)
PASS cssRule.type is 1
PASS declaration.length is 0
PASS declaration.getPropertyValue('box-shadow') is null
PASS successfullyParsed is true

TEST COMPLETE
